Refactor: Add plugins/common-components
